+++ src/main/kotlin/top/goopper/platform/dto/course/create/CreateSectionDTO.kt
@@ -1,8 +1,8 @@
 package top.goopper.platform.dto.course.create
 
 data class CreateSectionDTO(
-    val id: Long?,
-    var courseId: Long?,
+    val id: Int?,
+    var courseId: Int?,
     val name: String,
     val desc: String,
 )
