+++ src/main/kotlin/top/goopper/platform/dto/course/create/CreateTaskDTO.kt
@@ -3,10 +3,10 @@ package top.goopper.platform.dto.course.create
 import top.goopper.platform.dto.AttachmentDTO
 
 data class CreateTaskDTO(
-    val id: Long?,
-    var sectionId: Long?,
+    val id: Int?,
+    var sectionId: Int?,
     val name: String,
     val content: String,
     var attachments: List<AttachmentDTO>,
-    val submitTypeId: Long,
+    val submitTypeId: Int,
 )
