+++ src/main/kotlin/top/goopper/platform/dto/course/detail/TaskDetailDTO.kt
@@ -3,10 +3,10 @@ package top.goopper.platform.dto.course.detail
 import top.goopper.platform.dto.AttachmentDTO
 
 data class TaskDetailDTO(
-    val id: Long,
+    val id: Int,
     val name: String,
     val content: String,
     var status: Boolean,
-    var score: Long,
+    var score: Int,
     var attachment: List<AttachmentDTO>,
 )
